# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: DefterX API
  version: 1.0.0
  description: Offline-first finance API; monetary JSON properties are decimal strings.
servers:
  - url: /api/v1
paths:
...
  /reports/analytics: { get: { summary: Five-report analytics suite with trend, account history, category drill-down, liquidity forecast and investment performance, parameters: [{ in: query, name: from, schema: { type: string, format: date-time } }, { in: query, name: to, schema: { type: string, format: date-time } }, { in: query, name: accountIds, description: Comma-separated account UUIDs; omit for all or use none for no accounts, schema: { type: string } }, { in: query, name: granularity, schema: { type: string, enum: [day, week, month, year] } }], responses: { '200': { description: Analytics datasets }, '422': { description: Invalid range or account filter } } } }
  /reports/income-expense: { get: { summary: Income/expense with cost-center breakdown, parameters: [{ in: query, name: from, schema: { type: string, format: date-time } }, { in: query, name: to, schema: { type: string, format: date-time } }, { in: query, name: accountIds, description: Comma-separated account UUIDs; omit for all or use none for no accounts, schema: { type: string } }], responses: { '200': { description: Report } } } }
...